Move token function into own file
